backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-docx==1.1.0
python-multipart==0.0.9